+++ public/scripts/modules/sample-module/main.js
@@ -1,4 +1,0 @@
-define(['./sample-module', './sample-directive', './sample-filter', './sample-service', './dashboard-controller',
-    './sample-controller', './predix-asset-service', './predix-user-service', './predix-view-service'], function() {
-
-});