+++ public/scripts/modules/sample-module/sample-service.js
@@ -1,7 +1,0 @@
-define(['angular', './sample-module'], function(angular, sampleModule) {
-    'use strict';
-
-    sampleModule.value('version', '0.1');
-
-    return sampleModule;
-});