+++ tasks/compile.index.js
@@ -1,0 +1,16 @@
+'use strict';
+
+// -------------------------------------
+//   Task: Compile: Inline Index Source
+// -------------------------------------
+const inlinesource = require('gulp-inline-source');
+const rename = require('gulp-rename');
+
+module.exports = function(gulp) {
+  return function() {
+    gulp.src('./public/_index.html')
+      .pipe(inlinesource())
+      .pipe(rename('index.html'))
+      .pipe(gulp.dest('./public'));
+  };
+};