+++ tasks/serve.dev.start.js
@@ -1,0 +1,18 @@
+'use strict';
+
+// -------------------------------------
+//   Task: Serve
+// -------------------------------------
+const nodemon = require('gulp-nodemon');
+
+module.exports = function() {
+  return function() {
+    nodemon({
+        script: 'server/app.js',
+        env: { 'base-dir' : '/../public'}
+      })
+      .on('restart', function() {
+        console.log('app.js restarted');
+      });
+  };
+};