+++ HISTORY.md
@@ -1,3 +1,11 @@
+v1.4.2
+===================
+* Fix dist build issue stalling on promise polyfill.
+
+v1.4.1
+===================
+* Upgrade to Polymer 1.5.0 and latest dev dependencies.
+
 v1.4.0
 ===================
 * Upgrade to Polymer 1.4.0